requirements.txt
aiogram==3.15.0
psutil==6.1.0
python-dotenv==1.0.1
aiohttp==3.10.10

